+++ generators/_base/BaseSubGenerator.js
@@ -1,3 +1,4 @@
+const ejs = require('ejs');
 const path = require('path');
 
 const BaseGenerator = require('./BaseGenerator');
@@ -15,10 +16,20 @@ module.exports = class BaseSubGenerator extends BaseGenerator {
     // Calling the super constructor is important so our generator is correctly set up
     super(args, opts);
 
-    this.option('install', { type: Boolean, default: false, description: 'Install dependencies for generated code' });
+    this.option('install', {
+      type: Boolean,
+      default: false,
+      description: 'Install dependencies for generated code',
+    });
   }
 
-  transformName(name) { return name; }
+  transformName(name) {
+    return name;
+  }
+
+  transformModuleName(name) {
+    return name;
+  }
 
   mainArgument(name, description) {
     this._originalName = name;
@@ -26,7 +37,7 @@ module.exports = class BaseSubGenerator extends BaseGenerator {
     this.argument(this._name, {
       type: String,
       description,
-      required: true
+      required: true,
     });
   }
 
@@ -53,8 +64,10 @@ module.exports = class BaseSubGenerator extends BaseGenerator {
   }
 
   _checkMainArgumentName(name) {
-    if (!(/^[A-Z][a-zA-Z\d]*$/.test(name))) {
-      throw new Error(`Last name's part should start with Capital latin letter, and may includes only latin letters and numbers. You passed "${name}"`);
+    if (!/^[A-Z][a-zA-Z\d]*$/.test(name)) {
+      throw new Error(
+        `Last name's part should start with Capital latin letter, and may includes only latin letters and numbers. You passed "${name}"`,
+      );
     }
   }
 
@@ -72,34 +85,45 @@ module.exports = class BaseSubGenerator extends BaseGenerator {
 
   _calcDest() {
     this.options.dest = [
-      ...this.options.featureName.split('/').filter(Boolean).map(part => `features/${part}`),
-      this._dest,
-      this.options.Name,
+      ...this.options.featureName
+        .split('/')
+        .filter(Boolean)
+        .map(part => `features/${part}`),
+      this._dest ? `${this._dest}/${this.options.Name}` : '',
     ].join('/');
   }
 
   _calcModuleName() {
-    this.options.moduleName = [
-      this.options.FeatureName,
-      pascalize(this._originalName),
-      this.options.Name,
-    ].filter(Boolean).join('/');
+    this.options.moduleName = this.transformModuleName(
+      [this.options.FeatureName, pascalize(this._originalName), this.options.Name]
+        .filter(Boolean)
+        .join('/'),
+    );
   }
 
-  install(...rest){
-    if (this.options['install']) {
+  install(...rest) {
+    if (this.options.install) {
       this.npmInstall(...rest);
     }
   }
 
   writeNamedTemplates({ files = [], units = [], stories = [] }) {
-    const { name, Name, moduleName, featureName, FeatureName, flow, unit: writeUnit, stories: writeStories, dest } = this.options;
+    const {
+      name,
+      Name,
+      moduleName,
+      featureName,
+      FeatureName,
+      flow,
+      unit: writeUnit,
+      stories: writeStories,
+      dest,
+    } = this.options;
 
     const filesToWrite = [...files];
 
     if (writeUnit) filesToWrite.push(...units);
     if (writeStories) filesToWrite.push(...stories);
-
 
     filesToWrite.forEach(file => {
       const [realFileName, ...paths] = file.split('/').reverse();
@@ -108,6 +132,148 @@ module.exports = class BaseSubGenerator extends BaseGenerator {
         this.templatePath(`${file}.ejs`),
         this.destinationPath(path.join(dest, ...paths.reverse(), `${Name}.${realFileName}`)),
         { name, Name, featureName, moduleName, FeatureName, flow },
+      );
+    });
+  }
+
+  writeMissedTemplates({ files = [], units = [], stories = [] }) {
+    const {
+      name,
+      Name,
+      moduleName,
+      featureName,
+      FeatureName,
+      flow,
+      unit: writeUnit,
+      stories: writeStories,
+      dest,
+    } = this.options;
+
+    const filesToWrite = [...files];
+
+    if (writeUnit) filesToWrite.push(...units);
+    if (writeStories) filesToWrite.push(...stories);
+
+    filesToWrite.forEach(file => {
+      const [realFileName, ...paths] = file.split('/').reverse();
+
+      const pathToRealFile = path.join(dest, ...paths.reverse(), realFileName);
+
+      if (this.fs.exists(this.destinationPath(pathToRealFile))) return;
+
+      this.fs.copyTpl(this.templatePath(`${file}.main.ejs`), this.destinationPath(pathToRealFile), {
+        name,
+        Name,
+        featureName,
+        moduleName,
+        FeatureName,
+        flow,
+      });
+    });
+  }
+
+  writeMissedNamedTemplates({ files = [], units = [], stories = [] }) {
+    const {
+      name,
+      Name,
+      moduleName,
+      featureName,
+      FeatureName,
+      flow,
+      unit: writeUnit,
+      stories: writeStories,
+      dest,
+    } = this.options;
+
+    const filesToWrite = [...files];
+
+    if (writeUnit) filesToWrite.push(...units);
+    if (writeStories) filesToWrite.push(...stories);
+
+    filesToWrite.forEach(file => {
+      const [realFileName, ...paths] = file.split('/').reverse();
+
+      const pathToRealFile = path.join(dest, ...paths.reverse(), `${Name}.${realFileName}`);
+
+      if (this.fs.exists(this.destinationPath(pathToRealFile))) return;
+
+      this.fs.copyTpl(this.templatePath(`${file}.main.ejs`), this.destinationPath(pathToRealFile), {
+        name,
+        Name,
+        featureName,
+        moduleName,
+        FeatureName,
+        flow,
+      });
+    });
+  }
+
+  appendTemplates({ files = [], units = [], stories = [] }) {
+    const {
+      name,
+      Name,
+      moduleName,
+      featureName,
+      FeatureName,
+      flow,
+      unit: writeUnit,
+      stories: writeStories,
+      dest,
+    } = this.options;
+
+    const filesToWrite = [...files];
+
+    if (writeUnit) filesToWrite.push(...units);
+    if (writeStories) filesToWrite.push(...stories);
+
+    filesToWrite.forEach(file => {
+      const [realFileName, ...paths] = file.split('/').reverse();
+
+      this.fs.append(
+        this.destinationPath(path.join(dest, ...paths.reverse(), realFileName)),
+        ejs.render(this.fs.read(this.templatePath(`${file}.part.ejs`)), {
+          name,
+          Name,
+          featureName,
+          moduleName,
+          FeatureName,
+          flow,
+        }),
+      );
+    });
+  }
+
+  appendNamedTemplates({ files = [], units = [], stories = [] }) {
+    const {
+      name,
+      Name,
+      moduleName,
+      featureName,
+      FeatureName,
+      flow,
+      unit: writeUnit,
+      stories: writeStories,
+      dest,
+    } = this.options;
+
+    const filesToWrite = [...files];
+
+    if (writeUnit) filesToWrite.push(...units);
+    if (writeStories) filesToWrite.push(...stories);
+
+    filesToWrite.forEach(file => {
+      const [realFileName, ...paths] = file.split('/').reverse();
+
+      this.fs.append(
+        this.destinationPath(path.join(dest, ...paths.reverse(), `${Name}.${realFileName}`)),
+        `\n${ejs.render(this.fs.read(this.templatePath(`${file}.part.ejs`)), {
+          name,
+          Name,
+          featureName,
+          moduleName,
+          FeatureName,
+          flow,
+        })}`,
       );
     });
   }