+++ generators/feature/index.js
@@ -3,10 +3,24 @@ const BaseSubGenerator = require('../_base/BaseSubGenerator');
 const DESTINATION_FOLDER = 'features';
 
 const TEMPLATES = {
-  files: ['components/ExampleHeader.js', 'links/index.js', 'pages/ExampleRootPage.js', 'pages/ExampleNestedPage.js', 'routes/index.js', 'index.js'],
+  files: [
+    'components/ExampleHeader.js',
+    'links/index.js',
+    'pages/ExampleRootPage.js',
+    'pages/ExampleNestedPage.js',
+    'routes/index.js',
+    'index.js',
+  ],
 };
 const NAMED_TEMPLATES = {
-  files: ['documentation.yml', 'links/root.link.js', 'links/nested.link.js', 'routes/paths.js', 'routes/routeStore.js', 'inject.js'],
+  files: [
+    'documentation.yml',
+    'links/root.link.js',
+    'links/nested.link.js',
+    'routes/paths.js',
+    'routes/routeStore.js',
+    'inject.js',
+  ],
 };
 
 module.exports = class FeatureGenerator extends BaseSubGenerator {
@@ -23,7 +37,10 @@ module.exports = class FeatureGenerator extends BaseSubGenerator {
   // used to use lower-cased feature names in paths
   _calcDest() {
     this.options.dest = [
-      ...this.options.featureName.split('/').filter(Boolean).map(part => `features/${part}`),
+      ...this.options.featureName
+        .split('/')
+        .filter(Boolean)
+        .map(part => `features/${part}`),
       this._dest,
       this.options.name,
     ].join('/');
@@ -31,10 +48,9 @@ module.exports = class FeatureGenerator extends BaseSubGenerator {
 
   // used to use prevent extra "feature" inside module
   _calcModuleName() {
-    this.options.moduleName = [
-      this.options.FeatureName,
-      this.options.Name,
-    ].filter(Boolean).join('/');
+    this.options.moduleName = [this.options.FeatureName, this.options.Name]
+      .filter(Boolean)
+      .join('/');
   }
 
   initializing() {